backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.36
alembic==1.13.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pdfplumber==0.11.0
python-docx==1.1.2
reportlab==4.2.0
httpx==0.27.0
email-validator
bcrypt==4.0.1
requests
